requirements.txt
Flask
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
Python 3.13.7